GaugeBand gradientRatio should be double array #10

+++ src/main/java/gwt/material/design/amcharts/client/ui/chart/js/options/AmGaugeAxis.java
@@ -49,7 +49,7 @@ public class AmGaugeAxis extends AmAxisBase {
     public double bandAlpha;
 
     @JsProperty
-    public int[] bandGradientRatio;
+    public double[] bandGradientRatio;
 
     @JsProperty
     public double bandOutlineAlpha;
